[Refactor] #459 - myPageFeedViewModel 구현
